separate getApplicableloadEstimators into new file
