+++ src/autopas/selectors/tuningStrategy/RandomSearch.h
@@ -10,7 +10,7 @@
 #include <set>
 
 #include "TuningStrategyInterface.h"
-#include "autopas/containers/LoadEstimators.h"
+#include "autopas/containers/CompatibleLoadEstimators.h"
 #include "autopas/utils/ExceptionHandler.h"
 #include "autopas/utils/NumberSet.h"
