+++ src/autopas/selectors/tuningStrategy/SetSearchSpaceBasedTuningStrategy.h
@@ -7,8 +7,8 @@
 #pragma once
 
 #include "TuningStrategyInterface.h"
+#include "autopas/containers/CompatibleLoadEstimators.h"
 #include "autopas/containers/CompatibleTraversals.h"
-#include "autopas/containers/LoadEstimators.h"
 #include "autopas/utils/ExceptionHandler.h"
 
 namespace autopas {
